[STEP07-08] DB 설계 개선 및 쿼리 성능 최적화

+++ src/test/java/com/hhplus/hhplus_ecommerce/coupon/domain/UserCouponTest.java
@@ -135,7 +135,7 @@ class UserCouponTest {
         UserCoupon userCoupon = UserCoupon.issue(1L, coupon);
 
         // when
-        int discount = userCoupon.calculateDiscount(100000);
+        Long discount = userCoupon.calculateDiscount(100000L);
 
         // then
         assertThat(discount).isEqualTo(10000);  // 10% 할인
